unify honk recursive verifiers
